refactor: training status 공통화
